requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
asyncssh==2.20.0
atpublic==5.1
attrs==25.1.0
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
cryptography==44.0.2
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.8
dvc==3.59.1
dvc-data==3.16.9
dvc-gs==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.115.11
filelock==3.17.0
flatten-dict==0.4.2
flufl.lock==8.1.0
frozenlist==1.5.0
fsspec==2025.2.0
funcy==2.0
gcsfs==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.2
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grandalf==0.8
greenlet==3.1.1
gto==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
iterative-telemetry==0.0.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
# langchain==0.3.19
# langchain-core==0.3.40
# langchain-openai==0.3.7
# langchain-text-splitters==0.3.6
# langsmith==0.3.11
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
networkx==3.4.2
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.65.2
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pygit2==1.17.0
Pygments==2.19.1
pygtrie==2.5.0
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scmrepo==3.3.10
semver==3.0.4
setuptools==75.8.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sqltrie==0.11.2
starlette==0.46.0
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
yarl==1.18.3
zc.lockfile==3.0.post1
zstandard==0.23.0
pandas==2.1.4
numpy==1.26.1
scipy==1.13.1
scikit-learn==1.5.1
torch==2.2.0
transformers==4.35.0
nltk==3.9.1
joblib==1.4.2
mlflow==2.20.4
xgboost==1.7.6
gunicorn==23.0.0
tqdm==4.67.1
regex==2024.9.11
imblearn==0.0
textstat==0.7.5
spacy==3.8.4
en_core_web_sm==3.8.0
streamlit==1.40.1
bcrypt==3.2.0
pymongo==4.11.3
PyPDF2==3.0.1
python-docx==1.1.2
docx==0.2.4
google==3.0.0
google.genai==1.9.0
python-dotenv==0.21.0
pymupdf==1.25.5
mammoth==1.9.0
libmagic==1.0
python-magic-bin==0.4.14
psycopg2-binary==2.9.10
fastapi==0.115.12
fastapi.responses==0.2.1
supabase==2.15.0
uv==0.6.14